requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.55.3
slowapi==0.1.9
pydantic==2.9.2
python-dotenv==1.0.1
